Move components from /repo/compare to /components
