+++ web/src/SourcegraphWebApp.scss
@@ -256,6 +256,7 @@ body,
 @import './repo/blob/discussions/DiscussionsTree';
 @import './components/SearchResult';
 @import './components/SearchResultMatch';
+@import './components/FileDiffNode';
 @import '~@sourcegraph/react-loading-spinner/lib/LoadingSpinner.css';
 
 @import '../../shared/src/index';
