+++ web/src/components/DiffStat.scss
