+++ web/src/components/FileDiffHunks.tsx
@@ -5,16 +5,16 @@ import { isEqual } from 'lodash'
 import * as React from 'react'
 import { combineLatest, NEVER, Observable, of, Subject, Subscription } from 'rxjs'
 import { distinctUntilChanged, filter, map, switchMap } from 'rxjs/operators'
-import { ActionItemAction } from '../../../../shared/src/actions/ActionItem'
-import { DecorationMapByLine, groupDecorationsByLine } from '../../../../shared/src/api/client/services/decoration'
-import { HoverMerged } from '../../../../shared/src/api/client/types/hover'
-import { ExtensionsControllerProps } from '../../../../shared/src/extensions/controller'
-import * as GQL from '../../../../shared/src/graphql/schema'
-import { isDefined } from '../../../../shared/src/util/types'
-import { FileSpec, RepoSpec, ResolvedRevSpec, RevSpec, toURIWithPath } from '../../../../shared/src/util/url'
-import { ThemeProps } from '../../../../shared/src/theme'
+import { ActionItemAction } from '../../../shared/src/actions/ActionItem'
+import { DecorationMapByLine, groupDecorationsByLine } from '../../../shared/src/api/client/services/decoration'
+import { HoverMerged } from '../../../shared/src/api/client/types/hover'
+import { ExtensionsControllerProps } from '../../../shared/src/extensions/controller'
+import * as GQL from '../../../shared/src/graphql/schema'
+import { isDefined } from '../../../shared/src/util/types'
+import { FileSpec, RepoSpec, ResolvedRevSpec, RevSpec, toURIWithPath } from '../../../shared/src/util/url'
+import { ThemeProps } from '../../../shared/src/theme'
 import { DiffHunk } from './DiffHunk'
-import { diffDomFunctions } from './dom-functions'
+import { diffDomFunctions } from '../repo/compare/dom-functions'
 
 interface PartFileInfo {
     repoName: string
