+++ web/src/components/FileDiffNode.tsx
@@ -4,13 +4,13 @@ import ChevronDownIcon from 'mdi-react/ChevronDownIcon'
 import ChevronUpIcon from 'mdi-react/ChevronUpIcon'
 import * as React from 'react'
 import { Link } from 'react-router-dom'
-import { ActionItemAction } from '../../../../shared/src/actions/ActionItem'
-import { HoverMerged } from '../../../../shared/src/api/client/types/hover'
-import * as GQL from '../../../../shared/src/graphql/schema'
-import { FileSpec, RepoSpec, ResolvedRevSpec, RevSpec } from '../../../../shared/src/util/url'
+import { ActionItemAction } from '../../../shared/src/actions/ActionItem'
+import { HoverMerged } from '../../../shared/src/api/client/types/hover'
+import * as GQL from '../../../shared/src/graphql/schema'
+import { FileSpec, RepoSpec, ResolvedRevSpec, RevSpec } from '../../../shared/src/util/url'
 import { DiffStat } from './DiffStat'
 import { FileDiffHunks } from './FileDiffHunks'
-import { ThemeProps } from '../../../../shared/src/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 
 export interface FileDiffNodeProps extends ThemeProps {
     node: GQL.IFileDiff | GQL.IPreviewFileDiff
