+++ web/src/enterprise/campaigns/detail/CampaignDetails.tsx
@@ -38,7 +38,7 @@ import { ThemeProps } from '../../../../../shared/src/theme'
 import { TabsWithLocalStorageViewStatePersistence } from '../../../../../shared/src/components/Tabs'
 import SourcePullIcon from 'mdi-react/SourcePullIcon'
 import { LinkOrSpan } from '../../../../../shared/src/components/LinkOrSpan'
-import { FileDiffNode } from '../../../repo/compare/FileDiffNode'
+import { FileDiffNode } from '../../../components/FileDiffNode'
 
 interface Props extends ThemeProps {
     /**
