+++ web/src/repo/compare/RepositoryCompareArea.scss
@@ -1,5 +1,3 @@
-@import './FileDiffNode';
-
 .repository-compare-area {
     position: relative; // for tooltip positioning
 }
